Update importingdata.html
The one before this works with two places. This one adds 2 more salons.  I really want to run these out of a database but I'm too lazy to learn how that stuff works. I also want to figure out how to display the information in a side bar (including price, which is not a property)